requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
django-taggit==0.22.2
gunicorn==19.8.1
idna==2.6
omdb==0.10.0
psycopg2==2.7.4
python-http-client==3.0.0
pytz==2018.4
requests==2.18.4
sendgrid==5.3.0
urllib3==1.22
whitenoise==3.3.1
